@@ -3,40 +3,38 @@
   "version": "4.0.0-alpha.16",
   "private": true,
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "test": "pnpm run test:e2e && kill-port 4042",
-    "test:e2e": "start-server-and-test api 'http-get://localhost:4042/graphql?query=%7B__typename%7D' test:e2e:client",
-    "test:e2e:client": "vue-cli-service test:e2e --mode production --headless",
-    "test:e2e:dev": "start-server-and-test api:dev 'http-get://localhost:4042/graphql?query=%7B__typename%7D' test:e2e:dev:client",
-    "test:e2e:dev:client": "vue-cli-service test:e2e --mode development",
+    "dev": "vite --port 8080",
+    "build": "vite build",
+    "preview": "vite preview --port 8080",
+    "test": "pnpm run test:e2e",
+    "test:e2e": "start-server-and-test preview http://localhost:8080 test:e2e:run",
+    "test:e2e:run": "start-server-and-test api 'http-get://localhost:4042/graphql?query=%7B__typename%7D' test:e2e:cy",
+    "test:e2e:cy": "cypress run --headless",
+    "test:e2e:dev": "cypress open",
     "api": "test-server --simulate-latency 50",
-    "api:dev": "test-server  --simulate-latency 500"
+    "api:dev": "test-server --simulate-latency 500"
   },
   "dependencies": {
-    "@apollo/client": "^3.7.7",
+    "@apollo/client": "^3.7.16",
     "@vue/apollo-composable": "workspace:*",
     "@vue/apollo-util": "workspace:*",
-    "core-js": "^3.23.2",
-    "graphql": "^16.6.0",
+    "graphql": "^16.7.1",
     "graphql-tag": "^2.12.6",
-    "graphql-ws": "^5.13.1",
-    "regenerator-runtime": "^0.13.9",
     "test-server": "workspace:*",
-    "vue": "^3.2.37",
-    "vue-demi": "^0.13.1",
-    "vue-router": "^4.0.16"
+    "vue": "^3.3.4",
+    "@vitejs/plugin-vue": "^4.2.3",
+    "vue-router": "^4.2.4"
   },
   "devDependencies": {
-    "@babel/core": "^7.18.5",
-    "@vue/cli-plugin-babel": "^5.0.6",
-    "@vue/cli-plugin-e2e-cypress": "^5.0.6",
-    "@vue/cli-plugin-typescript": "^5.0.6",
-    "@vue/cli-service": "^5.0.6",
-    "axios": "^0.24.0",
-    "cypress": "^12.14.0",
-    "kill-port": "^1.6.1",
-    "start-server-and-test": "^1.14.0",
-    "tailwindcss": "^1.9.6"
+    "vite": "^4.4.2",
+    "vue-tsc": "^1.8.3",
+    "typescript": "^5.0.2",
+    "cypress-vite": "^1.4.1",
+    "axios": "^1.4.0",
+    "cypress": "^12.17.0",
+    "start-server-and-test": "^2.0.0",
+    "tailwindcss": "^3.3.2",
+    "postcss": "^8.4.25",
+    "autoprefixer": "^10.4.14"
   }
 }
