@@ -61,13 +61,13 @@
     }
   },
   "devDependencies": {
-    "@apollo/client": "^3.7.7",
+    "@apollo/client": "^3.7.16",
     "@types/throttle-debounce": "^5.0.0",
-    "graphql": "^16.6.0",
+    "graphql": "^16.7.1",
     "graphql-tag": "^2.12.6",
     "nodemon": "^1.19.4",
-    "rimraf": "^3.0.2",
-    "typescript": "^4.7.4",
-    "vue": "^3.2.37"
+    "rimraf": "^5.0.1",
+    "typescript": "^4.9.5",
+    "vue": "^3.3.4"
   }
 }
