@@ -1,8 +1,6 @@
-/* eslint-disable */
-
 module.exports = {
-  plugins: [
-    require('tailwindcss')(),
-    require('autoprefixer')(),
-  ],
+  plugins: {
+    tailwindcss: {},
+    autoprefixer: {},
+  },
 }
