@@ -3,7 +3,7 @@ import { DefaultApolloClient } from '@vue/apollo-composable'
 import { apolloClient } from './apollo'
 import App from './components/App.vue'
 import { router } from './router'
-import '@/assets/styles/tailwind.postcss'
+import '@/assets/styles/tailwind.css'
 
 const app = createApp({
   setup () {
